requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==4.4.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.2.5
pillow==10.4.0
proglog==0.1.11
python-dotenv==1.1.0
requests==2.32.3
tqdm==4.67.1
urllib3==2.4.0
